Init Grammatical Tools
GrammaticalTools.py is a special file that contains tools for converting text into the correct form. The English language is contain limited variants of grammatical cases, so simply adding "Add" at the beginning of a line makes the following text not grammatically correct. Of course, is possible to duplicate these variants of translations in * .ts files, but this will greatly increase it. This simple and short script allows you to avoid this. At a minimum, this script removes the second capital letters. Perhaps this is a temporary solution and in the future something better will be invented. Now this is worked only for the Russian language and this will not affect to other languages ​​in any way. However, if desired, other developers or translators can add their code to this script for own language.