WIP: Rename GUI commands to follow FreeCAD format
FreeCAD have a Help module where using the What's this? command you can click on a command on the toolbar and
the wiki page for it will open.The way it works is: it opens https://wiki.freecad.org/{$Command_name}.

For Fasteners WB pages are "Fasteners_CommandName", currently command names are "FSCommandName" which leads to an
error 404 when fetching Wiki page.

Fix #33